% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rnaedit_df.R
\docType{data}
\name{rnaedit_df}
\alias{rnaedit_df}
\title{Example breast cancer RNA editing dataset.}
\format{
A data frame containing RNA editing levels for 272 sites (in the 
  rows) for 221 subjects (in the columns). Row names are site IDs and column 
  names are sample IDs.
}
\source{
Synapse database ID: syn2374375.
}
\usage{
rnaedit_df
}
\description{
A subset of the TCGA breast cancer RNA editing dataset for 272
  edited sites on genes PHACTR4, CCR5, METTL7A and a few randomly sampled
  sites for 221 subjects.
}
\keyword{datasets}
